Add isTopLevel boolean property to Struct model; update StructConverter with method to determine; update CCodeGenerator for new requirement.
Update generateArrayMemberParser to match refactored C SDK flow for an array_parser

Modify RequestHelper to fix the parameterCheckingBlock for requests that have to params, such as getService()

Update Request templates to flow from ds3_c template, and prevent multiple generation.

Update a few .xml test input files to have the related Request for the Structs to be able to verify isTopLevel property.